+++ AviaNZ_batch_GUI.py
@@ -74,11 +74,11 @@ class AviaNZ_batchWindow(QMainWindow):
         self.area.addDock(self.d_detection, 'right')
         self.area.addDock(self.d_files, 'left')
 
-        self.w_browse = QPushButton("  &Browse Folder")
-        self.w_browse.setToolTip("Can select a folder with sub folders to process")
+        self.w_browse = QPushButton("  Browse Folder")
+        self.w_browse.setToolTip("Select a folder to process (may contain sub folders)")
         self.w_browse.setFixedSize(165, 50)
         self.w_browse.setIcon(self.style().standardIcon(QtGui.QStyle.SP_DialogOpenButton))
-        self.w_browse.setStyleSheet('QPushButton {font-weight: bold; font-size:14px; padding: 3px 3px 3px 3px}')
+        self.w_browse.setStyleSheet('QPushButton {font-weight: bold; padding: 3px 3px 3px 3px}')
         self.w_dir = QPlainTextEdit()
         self.w_dir.setFixedHeight(50)
         self.w_dir.setReadOnly(True)
@@ -135,11 +135,10 @@ class AviaNZ_batchWindow(QMainWindow):
         self.maxgap.valueChanged.connect(self.maxGapChange)
         self.maxgaplbl = QLabel("Maximum gap between syllables: 1 sec")
 
-        self.w_processButton = QPushButton(" &Process Folder")
-        self.w_processButton.setStyleSheet('QPushButton {font-weight: bold; font-size:14px; padding: 2px 2px 2px 8px}')
+        self.w_processButton = SupportClasses_GUI.MainPushButton(" Process Folder")
         self.w_processButton.setIcon(QIcon(QPixmap('img/process.png')))
         self.w_processButton.clicked.connect(self.detect)
-        self.w_processButton.setFixedSize(165, 50)
+        self.w_processButton.setFixedWidth(165)
         self.w_processButton.setEnabled(False)
         self.w_browse.clicked.connect(self.browse)
 
@@ -202,7 +201,6 @@ class AviaNZ_batchWindow(QMainWindow):
         colourPossibleDark = QColor(self.batchProc.config['ColourPossible'][0], self.batchProc.config['ColourPossible'][1], self.batchProc.config['ColourPossible'][2], 255)
         colourNamed = QColor(self.batchProc.config['ColourNamed'][0], self.batchProc.config['ColourNamed'][1], self.batchProc.config['ColourNamed'][2], self.batchProc.config['ColourNamed'][3])
         self.listFiles = SupportClasses_GUI.LightedFileList(colourNone, colourPossibleDark, colourNamed)
-        self.listFiles.setMinimumWidth(150)
         self.listFiles.itemDoubleClicked.connect(self.listLoadFile)
 
         self.w_files.addWidget(QLabel('Double click to select a folder'), row=0, col=0)
@@ -404,7 +402,7 @@ class AviaNZ_batchWindow(QMainWindow):
 
     def fillSpeciesBoxes(self):
         # select filters with Fs matching box 1 selection
-        # and show/hide minlen maxgap sliders
+        # and show/hide any other UI elements specific to bird filters or AnySound methods
         spp = []
         currname = self.w_spe1.currentText()
         if currname not in ["Any sound", "Any sound (Intermittent sampling)"]:
@@ -440,6 +438,18 @@ class AviaNZ_batchWindow(QMainWindow):
             self.boxTime.hide()
             self.addSp.hide()
             self.warning.show()
+
+        if currname=="NZ Bats":
+            self.addSp.setEnabled(False)
+            self.addSp.setToolTip("Bat recognisers cannot be combined with others")
+            self.w_wind.setChecked(False)
+            self.w_wind.setEnabled(False)
+            self.w_wind.setToolTip("Filter not applicable to bats")
+        else:
+            self.addSp.setEnabled(True)
+            self.addSp.setToolTip("")
+            self.w_wind.setEnabled(True)
+            self.w_wind.setToolTip("")
 
         # (skip first box which is fixed)
         for box in self.speCombos[1:]:
@@ -530,14 +540,14 @@ class AviaNZ_reviewAll(QMainWindow):
         # Make the window and set its size
         self.area = DockArea()
         self.setCentralWidget(self.area)
-        self.setMinimumSize(900, 700)
+        self.setMinimumSize(1000, 750)
         self.setWindowIcon(QIcon('img/Avianz.ico'))
 
         # Make the docks
         self.d_detection = Dock("Review",size=(600, 250), autoOrientation=False)
-        self.d_files = Dock("File list", size=(300, 700))
+        self.d_files = Dock("File list", size=(300, 750))
         self.d_excel = Dock("Excel", size=(600, 150))
-        self.d_settings = Dock("Advanced settings", size=(600, 300))
+        self.d_settings = Dock("Advanced settings", size=(600, 350))
         self.d_excel.hideTitleBar()
         self.d_settings.hideTitleBar()
 
@@ -548,32 +558,28 @@ class AviaNZ_reviewAll(QMainWindow):
 
         self.w_revLabel = QLabel("Reviewer")
         self.w_reviewer = QLineEdit()
-        self.w_browse = QPushButton("  &Browse Folder")
-        self.w_browse.setToolTip("Can select a folder with sub folders to process")
+        self.w_browse = QPushButton("  Browse Folder")
+        self.w_browse.setToolTip("Select a folder to review (may contain sub folders)")
         self.w_browse.setFixedHeight(50)
-        self.w_browse.setStyleSheet('QPushButton {font-weight: bold; font-size:14px}')
+        self.w_browse.setStyleSheet('QPushButton {font-weight: bold}')
         self.w_browse.setIcon(self.style().standardIcon(QtGui.QStyle.SP_DialogOpenButton))
         self.w_dir = QPlainTextEdit()
         self.w_dir.setFixedHeight(50)
         self.w_dir.setPlainText('')
         self.w_dir.setToolTip("The folder being processed")
 
-        self.w_processButton = QPushButton(" Review One-By-One")
-        self.w_processButton.setStyleSheet('QPushButton {font-weight: bold; font-size:14px; padding: 3px 3px 3px 8px}')
-        self.w_processButton.setFixedHeight(45)
-        self.w_processButton.setFixedHeight(45)
+        self.w_processButton = SupportClasses_GUI.MainPushButton(" Review One-By-One")
         self.w_processButton.setIcon(QIcon(QPixmap('img/review.png')))
         self.w_processButton.clicked.connect(self.reviewClickedAll)
         self.w_processButton.setEnabled(False)
-        self.w_processButton1 = QPushButton(" Review Quick")
-        self.w_processButton1.setStyleSheet('QPushButton {font-weight: bold; font-size:14px; padding: 3px 3px 3px 8px}')
-        self.w_processButton1.setFixedHeight(45)
-        self.w_processButton1.setFixedHeight(45)
+        self.w_processButton1 = SupportClasses_GUI.MainPushButton(" Review Quick")
         self.w_processButton1.setIcon(QIcon(QPixmap('img/tile1.png')))
         self.w_processButton1.clicked.connect(self.reviewClickedSingle)
         self.w_processButton1.setEnabled(False)
+        self.w_processButton.setMinimumWidth(200)
+        self.w_processButton1.setMinimumWidth(200)
 
-        self.w_speLabel1 = QLabel("Choose a species (or review all):")
+        self.w_speLabel1 = QLabel("Species to review")
         self.w_spe1 = QComboBox()
         self.w_spe1.currentIndexChanged.connect(self.speChanged)
         self.spList = []
@@ -581,34 +587,36 @@ class AviaNZ_reviewAll(QMainWindow):
         self.w_spe1.addItems(self.spList)
         self.w_spe1.setEnabled(False)
 
-        #minCertLab = QLabel("Skip if certainty above:\n")
-        minCertLab = QLabel("Skip if certainty above:\nTo see all annotation set to 100")
-        #minCertExp = QLabel("To see all annotation set to 100")
-        self.certBox = QSpinBox()
-        self.certBox.setRange(0,100)
-        self.certBox.setSingleStep(10)
-        self.certBox.setValue(90)
+        # Simple certainty selector:
+        self.certCombo = QComboBox()
+        self.certCombo.addItems(["Show all (even previously reviewed)", "Show only auto/unknown", "Custom certainty bounds"])
+        self.certCombo.setCurrentIndex(1)
+        self.certCombo.activated.connect(self.changedCertSimple)
 
         # add controls to dock
-        self.d_detection.addWidget(self.w_revLabel, row=0, col=0)
-        self.d_detection.addWidget(self.w_reviewer, row=0, col=1, colspan=2)
-        self.d_detection.addWidget(self.w_dir, row=1,col=1,colspan=2)
-        self.d_detection.addWidget(self.w_browse, row=1,col=0)
+        self.d_detection.addWidget(self.w_dir, row=0,col=1, colspan=2)
+        self.d_detection.addWidget(self.w_browse, row=0,col=0)
+        self.d_detection.addWidget(self.w_revLabel, row=1, col=0)
+        self.d_detection.addWidget(self.w_reviewer, row=1, col=1, colspan=2)
         self.d_detection.addWidget(self.w_speLabel1,row=2,col=0)
-        self.d_detection.addWidget(self.w_spe1,row=2,col=1)
-        self.d_detection.addWidget(self.w_processButton1, row=2, col=2)
-        #self.d_detection.addWidget(allsplabel, row=3, col=0, colspan=2)
-        self.d_detection.addWidget(self.w_processButton, row=3, col=2)
-        self.d_detection.addWidget(minCertLab, row=3, col=0)
-        self.d_detection.addWidget(self.certBox, row=3, col=1)
-        #self.d_detection.addWidget(minCertExp, row=4, col=0)
+        self.d_detection.addWidget(self.w_spe1,row=2, col=1, colspan=2)
+
+        self.d_detection.addWidget(QLabel("Minimum certainty to show"), row=3, col=0)
+        self.d_detection.addWidget(self.certCombo, row=3, col=1, colspan=2)
+
+        # self.d_detection.addWidget(self.w_processButton1, row=4, col=1)
+        # self.d_detection.addWidget(self.w_processButton, row=4, col=2)
+        procBox = QHBoxLayout()
+        procBox.addStretch(5)
+        procBox.addWidget(self.w_processButton1)
+        procBox.addStretch(1)
+        procBox.addWidget(self.w_processButton)
+        procBox.addStretch(5)
+
+        self.d_detection.layout.addLayout(procBox, 4, 0, 1, 3)
 
         # Excel export section
-        linesep2 = QFrame()
-        linesep2.setFrameShape(QFrame.HLine)
-        linesep2.setFrameShadow(QFrame.Sunken)
-        #self.d_detection.addWidget(linesep2, row=5, col=0, colspan=3)
-        self.w_resLabel = QLabel("Size (s) of presence/absence\nwindows in the output")
+        self.w_resLabel = QLabel("Size(s) of presence/absence windows\nin the output")
         self.w_res = QSpinBox()
         self.w_res.setRange(1,600)
         self.w_res.setSingleStep(5)
@@ -617,20 +625,20 @@ class AviaNZ_reviewAll(QMainWindow):
         self.timePrecisionBox = QComboBox()
         self.timePrecisionBox.addItems(["Down to seconds", "Down to milliseconds"])
         self.d_excel.addWidget(self.w_resLabel, row=6, col=0)
-        self.d_excel.addWidget(self.w_res, row=6, col=1)
+        self.d_excel.addWidget(self.w_res, row=6, col=1, colspan=2)
         self.d_excel.addWidget(timePrecisionLabel, row=7, col=0)
-        self.d_excel.addWidget(self.timePrecisionBox, row=7, col=1)
+        self.d_excel.addWidget(self.timePrecisionBox, row=7, col=1, colspan=2)
 
-        self.w_excelButton = QPushButton(" Generate Excel  ")
+        self.w_excelButton = QPushButton(" Generate Excel ")
         self.w_excelButton.setStyleSheet('QPushButton {font-weight: bold; font-size:14px; padding: 2px 2px 2px 8px}')
         self.w_excelButton.setFixedHeight(45)
         self.w_excelButton.setIcon(QIcon(QPixmap('img/excel.png')))
         self.w_excelButton.clicked.connect(self.exportExcel)
         self.w_excelButton.setEnabled(False)
-        self.d_excel.addWidget(self.w_excelButton, row=6, col=2)
+        self.d_excel.addWidget(self.w_excelButton, row=8, col=2)
 
         self.toggleSettingsBtn = QPushButton(" Advanced settings ")
-        self.toggleSettingsBtn.setStyleSheet('QPushButton {font-weight: bold; font-size:12px; padding: 2px 2px 2px 4px}')
+        self.toggleSettingsBtn.setStyleSheet('QPushButton {font-weight: bold; padding: 2px 2px 2px 4px}')
         self.toggleSettingsBtn.setFixedHeight(32)
         self.toggleSettingsBtn.setIcon(QIcon(QPixmap('img/settingsmore.png')))
         self.toggleSettingsBtn.setIconSize(QSize(25, 17))
@@ -639,6 +647,15 @@ class AviaNZ_reviewAll(QMainWindow):
         # linesep = QFrame()
         # linesep.setFrameShape(QFrame.HLine)
         # linesep.setFrameShadow(QFrame.Sunken)
+
+        # ADVANCED SETTINGS:
+
+        # precise certainty bounds
+        self.certBox = QSpinBox()
+        self.certBox.setRange(0,100)
+        self.certBox.setSingleStep(10)
+        self.certBox.setValue(90)
+        self.certBox.valueChanged.connect(self.changedCert)
 
         # sliders to select min/max frequencies for ALL SPECIES only
         self.fLow = QSlider(Qt.Horizontal)
@@ -685,20 +702,22 @@ class AviaNZ_reviewAll(QMainWindow):
         self.chunksizeBox.setValue(10)
         self.chunksizeBox.setEnabled(False)
 
-
+        # Advanced Settings Layout
         self.d_settings.addWidget(self.toggleSettingsBtn, row=0, col=2, colspan=2, rowspan=1)
-        self.d_settings.addWidget(self.fLowcheck, row=1, col=0)
-        self.d_settings.addWidget(self.fLowtext, row=1, col=1)
-        self.d_settings.addWidget(self.fLow, row=1, col=2, colspan=2, rowspan=1)
-        self.d_settings.addWidget(self.fLowvalue, row=1, col=4)
-        self.d_settings.addWidget(self.fHighcheck, row=2, col=0)
-        self.d_settings.addWidget(self.fHightext, row=2, col=1)
-        self.d_settings.addWidget(self.fHigh, row=2, col=2, colspan=2, rowspan=1)
-        self.d_settings.addWidget(self.fHighvalue, row=2, col=4)
-        self.d_settings.addWidget(QLabel("FFT window size"), row=3, col=1)
-        self.d_settings.addWidget(self.winwidthBox, row=3, col=2)
-        self.d_settings.addWidget(QLabel("FFT hop size"), row=3, col=3)
-        self.d_settings.addWidget(self.incrBox, row=3, col=4)
+        self.d_settings.addWidget(QLabel("Skip if certainty above:"), row=1, col=0, colspan=2, rowspan=1)
+        self.d_settings.addWidget(self.certBox, row=1, col=2, colspan=2, rowspan=1)
+        self.d_settings.addWidget(self.fLowcheck, row=2, col=0)
+        self.d_settings.addWidget(self.fLowtext, row=2, col=1)
+        self.d_settings.addWidget(self.fLow, row=2, col=2, colspan=2, rowspan=1)
+        self.d_settings.addWidget(self.fLowvalue, row=2, col=4)
+        self.d_settings.addWidget(self.fHighcheck, row=3, col=0)
+        self.d_settings.addWidget(self.fHightext, row=3, col=1)
+        self.d_settings.addWidget(self.fHigh, row=3, col=2, colspan=2, rowspan=1)
+        self.d_settings.addWidget(self.fHighvalue, row=3, col=4)
+        self.d_settings.addWidget(QLabel("FFT window size"), row=4, col=1)
+        self.d_settings.addWidget(self.winwidthBox, row=4, col=2)
+        self.d_settings.addWidget(QLabel("FFT hop size"), row=4, col=3)
+        self.d_settings.addWidget(self.incrBox, row=4, col=4)
 
         self.d_settings.addWidget(self.chunksizeAuto, row=5, col=0, colspan=2, rowspan=1)
         self.d_settings.addWidget(self.chunksizeManual, row=6, col=0, colspan=2, rowspan=1)
@@ -716,7 +735,6 @@ class AviaNZ_reviewAll(QMainWindow):
         colourPossibleDark = QColor(self.config['ColourPossible'][0], self.config['ColourPossible'][1], self.config['ColourPossible'][2], 255)
         colourNamed = QColor(self.config['ColourNamed'][0], self.config['ColourNamed'][1], self.config['ColourNamed'][2], self.config['ColourNamed'][3])
         self.listFiles = SupportClasses_GUI.LightedFileList(colourNone, colourPossibleDark, colourNamed)
-        self.listFiles.setMinimumWidth(150)
         self.listFiles.itemDoubleClicked.connect(self.listLoadFile)
         self.w_files.addWidget(self.listFiles, row=2, col=0)
 
@@ -736,21 +754,54 @@ class AviaNZ_reviewAll(QMainWindow):
         self.d_settings.layout.setColumnStretch(2, 5)
         self.show()
 
-    def toggleSettings(self):
-        if self.d_settings.widgets[1].isVisible():
-            # self.d_settings.setVisible(False)
-            for item in self.d_settings.widgets:
-                if item!=self.toggleSettingsBtn:
-                    item.hide()
-            self.toggleSettingsBtn.setText(" Advanced settings ")
-            self.toggleSettingsBtn.setIcon(QIcon(QPixmap('img/settingsmore.png')))
+    def changedCertSimple(self, cert):
+        # update certainty spinbox (adv setting) when dropdown changed
+        if cert==0:
+            # Will show all annotations
+            self.certBox.setValue(100)
+        elif cert==1:
+            # Will show yellow, red annotations
+            self.certBox.setValue(90)
         else:
+            # Will show a custom range
+            # Make sure the advanced settings dock is visible, to make it obvious
+            # where to change this parameter
+            self.toggleSettings(None, forceOn=True)
+            self.certBox.setFocus()
+
+    def changedCert(self, cert):
+        # update certainty dropdown when advanced setting changed
+        if cert==100:
+            # "Show all"
+            self.certCombo.setCurrentIndex(0)
+        elif cert==90:
+            # "Show yellow + red"
+            self.certCombo.setCurrentIndex(1)
+        else:
+            # "custom"
+            self.certCombo.setCurrentIndex(2)
+
+    def toggleSettings(self, clicked, forceOn=None):
+        """ forceOn can be None to toggle, or True/False to force Visible/Hidden. """
+        if forceOn is None:
+            forceOn = not self.d_settings.widgets[1].isVisible()
+
+        if forceOn:
             for item in self.d_settings.widgets:
                 if item!=self.toggleSettingsBtn:
                     item.show()
             # self.d_settings.setVisible(True)
+            self.d_excel.hide()
             self.toggleSettingsBtn.setText(" Hide settings ")
             self.toggleSettingsBtn.setIcon(QIcon(QPixmap('img/settingsless.png')))
+        else:
+            # self.d_settings.setVisible(False)
+            for item in self.d_settings.widgets:
+                if item!=self.toggleSettingsBtn:
+                    item.hide()
+            self.d_excel.show()
+            self.toggleSettingsBtn.setText(" Advanced settings ")
+            self.toggleSettingsBtn.setIcon(QIcon(QPixmap('img/settingsmore.png')))
         self.repaint()
         QApplication.processEvents()
 
@@ -774,8 +825,10 @@ class AviaNZ_reviewAll(QMainWindow):
     def speChanged(self, value):
         if self.w_spe1.currentText() == "All species":
             self.w_processButton1.setEnabled(False)
+            self.w_processButton1.setToolTip("Only one species at a time can be reviewed in quick mode")
         else:
             self.w_processButton1.setEnabled(True)
+            self.w_processButton1.setToolTip("")
 
     def fLowChanged(self, value):
         self.fLowvalue.setText(str(int(value)))
@@ -1502,7 +1555,8 @@ class AviaNZ_reviewAll(QMainWindow):
                             sp.readWav(filename, off=x1, len=x2-x1, silent=segix>1)
 
                             # Filter the audiodata based on initial sliders
-                            sp.data = sp.ButterworthBandpass(sp.data, sp.sampleRate, minFreq, maxFreq)
+                            # sp.data = sp.ButterworthBandpass(sp.data, sp.sampleRate, minFreq, maxFreq)
+                            sp.data = sp.bandpassFilter(sp.data, sp.sampleRate, minFreq, maxFreq)
 
                             # Generate the spectrogram
                             _ = sp.spectrogram(window='Hann', sgType='Standard',mean_normalise=True, onesided=True,need_even=False)
@@ -1555,12 +1609,14 @@ class AviaNZ_reviewAll(QMainWindow):
                 print("Segment matches")
                 del self.origSeg[i]
             else:
-                oldlabel = self.origSeg[i][4] 
-                newlabel = self.segments[i][4] 
+                oldlabel = self.origSeg[i][4]
+                newlabel = self.segments[i][4]
                 if "-To Be Deleted-" in [lab["species"] for lab in newlabel]:
                     self.origSeg[i] = [self.origSeg[i], []]
                 else:
-                    if [lab["species"] for lab in oldlabel] != [lab["species"] for lab in newlabel] or [lab["calltype"] for lab in oldlabel] != [lab["calltype"] for lab in newlabel]:
+                    # Note that we have to use .get to allow unspecified calltype
+                    if [lab["species"] for lab in oldlabel] != [lab["species"] for lab in newlabel] or \
+                       [lab.get("calltype") for lab in oldlabel] != [lab.get("calltype") for lab in newlabel]:
                         self.origSeg[i] = [self.origSeg[i], newlabel]
 
         if len(self.origSeg)>0: